[REL] stock_picking_responsible to V11. [FIX] Separate classes on different files [FIX] Rename files [FIX] Update README
