+++ stock_picking_responsible/__init__.py
