+++ stock_picking_responsible/__manifest__.py
@@ -19,7 +19,7 @@
 ##############################################################################
 {
     'name': 'Stock Picking Responsible',
-    'version': '9.0.1.0.0',
+    'version': '11.0.1.0.0',
     'category': 'Warehouse Management',
     'sequence': 14,
     'summary': '',
@@ -32,12 +32,12 @@
         'sale_stock',
     ],
     'data': [
-        'views/stock_move_view.xml',
-        'views/stock_picking_view.xml',
+        'views/stock_move_views.xml',
+        'views/stock_picking_views.xml',
     ],
     'demo': [
     ],
-    'installable': False,
+    'installable': True,
     'auto_install': False,
     'application': False,
 }
