+++ stock_picking_responsible/models/stock_picking.py
@@ -2,5 +2,13 @@
 # For copyright and license notices, see __manifest__.py file in module root
 # directory
 ##############################################################################
-from . import stock_move
-from . import stock_picking
+from odoo import models, fields
+
+
+class StockPicking(models.Model):
+    _inherit = 'stock.picking'
+
+    user_id = fields.Many2one(
+        'res.users',
+        string="Responsible"
+    )
