--- stock_picking_responsible/views/stock_move_view.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0"?>
-<odoo>
-    <data>
-
-        <record id="view_move_tree" model="ir.ui.view">
-            <field name="name">stock.move.tree.inherit</field>
-            <field name="model">stock.move</field>
-            <field name="inherit_id" ref="stock.view_move_tree"/>
-            <field name="arch" type="xml">
-                <tree>
-                    <field name="picking_user_id"/>
-                </tree>
-            </field>
-        </record>
-
-        <record id="view_move_form" model="ir.ui.view">
-            <field name="name">stock.move.form.inherit</field>
-            <field name="model">stock.move</field>
-            <field name="inherit_id" ref="stock.view_move_form"/>
-            <field name="arch" type="xml">
-                <field name="picking_id" position="after">
-                    <field name="picking_user_id"/>
-                </field>
-            </field>
-        </record>
-        
-    </data>
-</odoo>
