+++ stock_picking_responsible/views/stock_picking_views.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0"?>
+<odoo>
+    <record id="view_picking_form" model="ir.ui.view">
+        <field name="name">stock.picking.form.inherit</field>
+        <field name="model">stock.picking</field>
+        <field name="inherit_id" ref="stock.view_picking_form"/>
+        <field name="arch" type="xml">
+            <field name="partner_id" position="after">
+                <field name="user_id" attrs="{'readonly': [('state', '=', 'done')]}"/>
+            </field>
+        </field>
+    </record>
+    
+    <record id="vpicktree" model="ir.ui.view">
+        <field name="name">stock.picking.form.inherit</field>
+        <field name="model">stock.picking</field>
+        <field name="inherit_id" ref="stock.vpicktree"/>
+        <field name="arch" type="xml">
+            <field name="partner_id" position="after">
+                <field name="user_id"/>
+            </field>
+        </field>
+    </record>
+    
+    <record id="view_picking_internal_search" model="ir.ui.view">
+        <field name="name">stock.picking.filter.inherit</field>
+        <field name="model">stock.picking</field>
+        <field name="inherit_id" ref="stock.view_picking_internal_search"/>
+        <field name="arch" type="xml">
+            <field name="partner_id" position="after">
+                <field name="user_id"/>
+            </field>
+            <group>
+                <filter string="Responsible" domain="[]" context="{'group_by':'user_id'}"/>
+            </group>
+        </field>
+    </record>
+</odoo>
